+++ changelog.md
@@ -24,6 +24,7 @@ For example:
 * Link to table field can now link rows in the same table. [#798](https://gitlab.com/bramw/baserow/-/issues/798)
 * Made it clearer that you're navigating to baserow.io when clicking the "Get a license"
   button.
+* `./dev.sh all_in_one_dev` now starts a hot reloading dev mode using the all-in-one image.
 
 ### Bug Fixes